feat: use innerHTML
