+++ deno/hi.html
@@ -1,0 +1,14 @@
+<h1>Hi, there!</h1>
+<p>Hello, it’s Max.</p>
+<p>
+	A <a href="/allô">bilingual</a> developer &amp;, designer in
+	<span data-id="803">cloudy</span>
+	London. Currently working at The Guardian.
+</p>
+<p>
+	Elegance, simplicity and accessibility guide
+	<a href="/works">my work</a>. I love building for the web
+	and making
+	<a href="https://github.com/mxdvl">open-source contributions</a>.
+</p>
+<p>Feels like 242K</p>
