+++ deno/template.html
@@ -37,20 +37,7 @@
 			</nav>
 		</header>
 		<main>
-			<h1>Hi, there!</h1>
-			<p>Hello, it’s Max.</p>
-			<p>
-				A <a href="/allô">bilingual</a> developer &amp;, designer in
-				<span data-id="803">cloudy</span>
-				London. Currently working at The Guardian.
-			</p>
-			<p>
-				Elegance, simplicity and accessibility guide
-				<a href="/works">my work</a>. I love building for the web
-				and making
-				<a href="https://github.com/mxdvl">open-source contributions</a>.
-			</p>
-			<p>Feels like 288.94K</p>
+			<slot>
 		</main>
 		<footer>
 			<p>
@@ -75,6 +62,7 @@
 					content</a>.
 			</p>
 			<p>© MXDVL 2022</p>
+			<ul></ul>
 		</footer>
 	</div>
 </body>
